requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# OpenAI & LLM
openai==1.10.0

# Vector Search & Embeddings
faiss-cpu==1.7.4
sentence-transformers==5.1.0
numpy==1.26.2

# PDF Processing
pypdf==4.0.0

# Web Scraping
httpx==0.27.0
selectolax==0.4.0

# Scheduling
APScheduler==3.10.4

# Azure (optional, for deployment)
# azure-storage-blob==12.19.0
# azure-identity==1.15.0
